+++ x/wasm/binding/cw_struct/cw_msg.go
@@ -22,6 +22,7 @@ type BindingMsg struct {
 	PegShift              *PegShift              `json:"peg_shift,omitempty"`
 	DepthShift            *DepthShift            `json:"depth_shift,omitempty"`
 	SetMarketEnabled      *SetMarketEnabled      `json:"set_market_enabled,omitempty"`
+	CreateMarket          *CreateMarket          `json:"create_market,omitempty"`
 
 	EditOracleParams *EditOracleParams `json:"edit_oracle_params,omitempty"`
 }
@@ -97,4 +98,44 @@ type InsuranceFundWithdraw struct {
 type SetMarketEnabled struct {
 	Pair    string `json:"pair"`
 	Enabled bool   `json:"enabled"`
+}
+
+type CreateMarket struct {
+	Pair         string        `json:"pair"`
+	PegMult      sdk.Dec       `json:"peg_mult,omitempty"`
+	SqrtDepth    sdk.Dec       `json:"sqrt_depth,omitempty"`
+	MarketParams *MarketParams `json:"market_params,omitempty"`
+}
+
+type MarketParams struct {
+	Pair    string
+	Enabled bool `json:"enabled,omitempty"`
+	// percentage that a single open or close position can alter the reserve
+	// amounts
+	PriceFluctuationLimitRatio sdk.Dec
+	// the minimum margin ratio which a user must maintain on this market
+	MaintenanceMarginRatio sdk.Dec `json:"maintenance_margin_ratio"`
+	// the maximum leverage a user is able to be taken on this market
+	MaxLeverage sdk.Dec `json:"max_leverage"`
+	// Latest cumulative premium fraction for a given pair.
+	// Calculated once per funding rate interval.
+	// A premium fraction is the difference between mark and index, divided by the
+	// number of payments per day. (mark - index) / # payments in a day
+	LatestCumulativePremiumFraction sdk.Dec `json:"latest_cumulative_premium_fraction"`
+	// the percentage of the notional given to the exchange when trading
+	ExchangeFeeRatio sdk.Dec `json:"exchange_fee_ratio"`
+	// the percentage of the notional transferred to the ecosystem fund when
+	// trading
+	EcosystemFundFeeRatio sdk.Dec `json:"ecosystem_fund_fee_ratio"`
+	// the percentage of liquidated position that will be
+	// given to out as a reward. Half of the liquidation fee is given to the
+	// liquidator, and the other half is given to the ecosystem fund.
+	LiquidationFeeRatio sdk.Dec `json:"liquidation_fee_ratio"`
+	// the portion of the position size we try to liquidate if the available
+	// margin is higher than liquidation fee
+	PartialLiquidationRatio sdk.Dec `json:"partial_liquidation_ratio"`
+	// specifies the interval on which the funding rate is updated
+	FundingRateEpochId string `json:"funding_rate_epoch_id,omitempty"`
+	// amount of time to look back for TWAP calculations
+	TwapLookbackWindow sdk.Int `json:"twap_lookback_window"`
 }