+++ x/wasm/binding/exec_perp.go
@@ -1,6 +1,8 @@
 package binding
 
 import (
+	"time"
+
 	wasmvmtypes "github.com/CosmWasm/wasmvm/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
@@ -188,4 +190,46 @@ func (exec *ExecutorPerp) SetMarketEnabled(
 	}
 
 	return exec.PerpV2.Admin().SetMarketEnabled(ctx, pair, cwMsg.Enabled)
+}
+
+func (exec *ExecutorPerp) CreateMarket(
+	cwMsg *cw_struct.CreateMarket, ctx sdk.Context,
+) (err error) {
+	if cwMsg == nil {
+		return wasmvmtypes.InvalidRequest{Err: "null msg"}
+	}
+
+	pair, err := asset.TryNewPair(cwMsg.Pair)
+	if err != nil {
+		return err
+	}
+
+	var market perpv2types.Market
+	if cwMsg.MarketParams == nil {
+		market = perpv2types.DefaultMarket(pair)
+	} else {
+		mp := cwMsg.MarketParams
+		market = perpv2types.Market{
+			Pair:                            pair,
+			Enabled:                         true,
+			PriceFluctuationLimitRatio:      mp.PriceFluctuationLimitRatio,
+			MaintenanceMarginRatio:          mp.MaintenanceMarginRatio,
+			MaxLeverage:                     mp.MaxLeverage,
+			LatestCumulativePremiumFraction: mp.LatestCumulativePremiumFraction,
+			ExchangeFeeRatio:                mp.ExchangeFeeRatio,
+			EcosystemFundFeeRatio:           mp.EcosystemFundFeeRatio,
+			LiquidationFeeRatio:             mp.LiquidationFeeRatio,
+			PartialLiquidationRatio:         mp.PartialLiquidationRatio,
+			FundingRateEpochId:              mp.FundingRateEpochId,
+			TwapLookbackWindow:              time.Duration(mp.TwapLookbackWindow.Int64()),
+			PrepaidBadDebt:                  sdk.NewCoin(pair.QuoteDenom(), sdk.ZeroInt()),
+		}
+	}
+
+	return exec.PerpV2.Admin().CreateMarket(ctx, perpv2keeper.ArgsCreateMarket{
+		Pair:            pair,
+		PriceMultiplier: cwMsg.PegMult,
+		SqrtDepth:       cwMsg.SqrtDepth,
+		Market:          &market,
+	})
 }